+++ pmp/src/main/java/com/qg/controller/FrontendController.java
@@ -3,6 +3,7 @@ package com.qg.controller;
 import cn.hutool.json.JSONUtil;
 import com.qg.domain.*;
 import com.qg.dto.FileUploadDTO;
+import com.qg.service.FileUploadService;
 import com.qg.service.FrontendBehaviorService;
 import com.qg.service.FrontendErrorService;
 import com.qg.service.FrontendPerformanceService;
@@ -21,6 +22,7 @@ import java.util.List;
 import java.util.Map;
 
 import static com.qg.domain.Code.INTERNAL_ERROR;
+import static com.qg.utils.FileUploadHandler.determineSubDirectory;
 import static com.qg.utils.FileUploadHandler.isValidFile;
 
 /**
@@ -43,6 +45,8 @@ public class FrontendController {
 
     @Autowired
     private FrontendBehaviorService frontendBehaviorService;
+    @Autowired
+    private FileUploadService fileUploadService;
 
 
 //    @PostMapping("/performance")
@@ -77,7 +81,7 @@ public class FrontendController {
 //    }
 
     @PostMapping("/{type}")
-    public Result getData(@RequestBody String data,@PathVariable String type) {
+    public Result getData(@RequestBody String data, @PathVariable String type) {
 
         switch (type) {
             case "performance":
@@ -101,105 +105,15 @@ public class FrontendController {
 
     @PostMapping("/formData")
     public Result getFile(@RequestParam String projectId, @RequestParam String timestamp, @RequestParam String version,
-                        @RequestParam String buildVersion, @RequestParam MultipartFile[] files,
-                        @RequestParam String [] jsFilenames, @RequestParam String fileHashes) {
+                          @RequestParam String buildVersion, @RequestParam MultipartFile[] files,
+                          @RequestParam String[] jsFilenames, @RequestParam String fileHashes) {
 
-        log.info("\n项目ID: " + projectId +
-                "\n时间戳: " + timestamp +
-                "\n版本: " + version +
-                "\n构建版本: " + buildVersion +
-                "\n文件数量: " + files.length +
-                "\nJS文件名: " + String.join(", ", jsFilenames) +
-                "\n文件哈希: " + fileHashes);
+        log.info("\n项目ID: {}\n时间戳: {}\n版本: {}\n构建版本: {}\n文件数量: {}\nJS文件名: {}\n文件哈希: {}", projectId, timestamp, version, buildVersion, files.length, String.join(", ", jsFilenames), fileHashes);
 
-        try {
-            // 参数校验
-            if (files.length == 0) {
-                log.warn("文件上传失败：未接收到任何文件");
-                return new Result(Code.BAD_REQUEST, "至少需要上传一个文件");
-            }
-
-            // 构建DTO
-            FileUploadDTO dto = new FileUploadDTO();
-            dto.setProjectId(projectId);
-            dto.setTimestamp(timestamp);
-            dto.setVersion(version);
-            dto.setBuildVersion(buildVersion);
-            dto.setJsFilenames(jsFilenames);
-            dto.setFileHashes(fileHashes);
-
-            // 处理文件上传
-            List<FileInfo> uploadedFiles = processFiles(files, dto);
-
-            // 返回结构化响应
-//            return new Result(Code.SUCCESS, "文件上传成功", new Map());
-
-            return null;
-        } catch (IOException e) {
-            log.error("文件存储失败{}", e.getMessage());
-            return new Result(Code.INTERNAL_ERROR, "文件存储失败: " + e.getMessage());
-        } catch (Exception e) {
-            log.error("系统处理异常{}", e.getMessage());
-            return new Result(Code.INTERNAL_ERROR, "系统处理异常");
-        }
+        return fileUploadService.uploadFile(projectId, timestamp
+                , version, buildVersion, files, jsFilenames, fileHashes);
     }
 
-    /**
-     * 处理文件存储逻辑
-     */
-    private List<FileInfo> processFiles(MultipartFile[] files, FileUploadDTO dto) throws IOException {
-        List<FileInfo> result = new ArrayList<>();
-
-        for (MultipartFile file : files) {
-            // 1. 文件类型校验
-            if (!isValidFile(file)) {
-                log.warn("跳过无效文件: {}", file.getOriginalFilename());
-                continue;
-            }
-
-            // 2. 确定存储目录
-            String subDir = determineSubDirectory(file);
-
-            // 3. 保存文件
-            String fileUrl = FileUploadHandler.saveFile(file, subDir);
-
-            // 4. 记录文件信息
-            result.add(new FileInfo(
-                    file.getOriginalFilename(),
-                    fileUrl,
-                    file.getSize(),
-                    subDir
-            ));
-
-            log.debug("文件保存成功: {} -> {}", file.getOriginalFilename(), fileUrl);
-        }
-
-        return result;
-    }
-
-
-
-    /**
-     * 根据文件类型确定存储目录
-     */
-    private String determineSubDirectory(MultipartFile file) {
-        if (FileUploadHandler.isValidImageFile(file)) {
-            return FileUploadHandler.IMAGE_DIR;
-        } else if (FileUploadHandler.isValidDocumentFile(file)) {
-            return FileUploadHandler.DOCUMENT_DIR;
-        } else {
-            return FileUploadHandler.INSTALL_DIR;
-        }
-    }
-
-    @Data
-    @AllArgsConstructor
-    private static class FileInfo {
-        private String originalName;
-        private String storedPath;
-        private long size;
-        private String fileType;
-    }
 
 
 }
